update create profile link
